RF: Split up and move tests of former plugins
